feat: Implemented zany zoo
Zany zoo is the motor for running muultiple busy beaver programs.
Currently it's only in single threaded mode, but it can be extended to
run multiple programs in parallel later. The code has been tested with
multiple busy beaver programs and works as expected, but more tests
are needed to ensure robustness.

The program is currently missing ignore values which will cause
the busy beaver program to fail.

Breaking changes: None

Resolves: None