feat: fixes and improvements busy beaver
Implemented a simplified command line interface for the busy beaver.
Currently, it only supportes piping file contents to the program. The
program will read the contents of the file, parse it, and execute the
busy beaver program for each line of input.

Adds more documentation to the busy beaver program.

Adds support for - in the transistions. This needs to improve later to
support verification of invalid transitions.

Breaking changes: None

Resolves: None